Add support for git subcommands
`git help` shows manpages for git subcommands, named of the form
`git-subcommand`.  Therefore, it's possible to use any man reader to
show git help pages.

This commit makes it so when CMD is "git", a second word is read and the
existence of the corresponding git help page is tested.  If it exists,
that is the man page that's shown; otherwise the man page for git itself
is shown.