Refactor for special terminals, add screen support
Not all terminals support the same command-line arguments as xterm.  It
would be nice to not leave those terminals out in the cold, so this
commit refactors the script to allow terminals that don't work like
xterm to have special commands run.

As an initial example of why this is beneficial, this commit also adds
support for GNU Screen, which takes over the current terminal to show a
man page.  Quitting the man reader returns to the command line exactly
as it was.  It even works in a plain TTY, so the part that bails out in
a TTY has been changed to check the TERMINAL first, and the bind is done
unconditionally.