Add "man" pseudo-terminal
Okay, I'll be honest, there might be a nicer way of doing this
configuration-wise, but for now, setting TERMINAL=man causes MANPGM to
be run directly.  This lets `man` run in the same terminal the command
line is in like with TERMINAL=screen, but without running GNU Screen to
do it.  In fact, I feel like this makes TERMINAL=screen unnecessary, so
maybe that should be removed before this branch is merged.