Configure through environment variables
Now the configuration for the _bash_help function can be set through
BASHELP_* environment variables (with the exception of BROWSER, which
hasn't been renamed because it's a common environment variable).
They're mostly the same as the old ones, but the terminal-geometry
option has been changed to a general terminal-options setting.

With this change, the _bash_help function doesn't create any non-local
variables, which is nice.