Work with & and ; chaining
Now bashelp removes everything to the last character in [|&;] and all
the whitespace that follows before reading the command.  This makes it
work with more types of command chaining.  It's all done with one call
to sed now, which is a bit more compact than having multiple variable
substitutions done like before.

Limitations: it doesn't understand the various things that would escape
the delimiting characters, so that's a bummer.  It probably can't be
correctly done in general with just a regular expression.